refactor: 멤버 라운드 대기 구조 잡기, 불필요한 주석, 메서드 제거, Todo주석 추가
