+++ app/src/main/java/com/stormers/storm/round/fragment/MemberWaitingFragment.kt
@@ -3,25 +3,19 @@ package com.stormers.storm.round.fragment
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.stormers.storm.R
 import com.stormers.storm.customview.StormButton
+import com.stormers.storm.network.BaseResponse
 import com.stormers.storm.network.RetrofitClient
 import com.stormers.storm.network.SocketClient
-import com.stormers.storm.project.adapter.ParticipatedProjectListAdapter
 import com.stormers.storm.project.base.BaseProjectWaitingActivity
-import com.stormers.storm.project.model.ResponseProjectUserListModel
 import com.stormers.storm.round.base.BaseWaitingFragment
+import com.stormers.storm.round.model.RoundEnterModel
+import com.stormers.storm.round.network.InterfaceRoundEnter
 import com.stormers.storm.round.network.InterfaceRoundUser
 import com.stormers.storm.user.ParticipantAdapter
-import com.stormers.storm.user.UserModel
-import com.stormers.storm.util.MarginDecoration
 import io.socket.emitter.Emitter
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.activity_main.view.*
 import kotlinx.android.synthetic.main.activity_round_setting.*
-import kotlinx.android.synthetic.main.layout_list_of_participant.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -39,58 +33,23 @@ class MemberWaitingFragment : BaseWaitingFragment(R.layout.fragment_round_settin
 
         activityButton = (activity as BaseProjectWaitingActivity).stormButton_ok_host_round_setting
 
-        setRecyclerView(view)
-
         activityButton.run {
             visibility = View.GONE
         }
 
         //Todo: 소켓으로 라운드 설정이 마쳐지는지 확인
-        //Todo: 라운드 설정이 마쳐지면 GET 라운드 정보 (roundIdx, roundPurpose, roundTime 등)
-        //Todo: 라운드 정보를 사용하여 POST 라운드 참여
-        //Todo: 참가한 사용자 목록 GET
+        //Todo: 라운드 설정이 마쳐지면 GET 라운드 정보 (roundIdx, roundPurpose, roundTime 등) //getRoundInfo() done
+        //Todo: 라운드 정보를 사용하여 POST 라운드 참여 done
+        //Todo: 참가한 사용자 목록 GET done
         //Todo: 소켓으로 라운드가 시작되는지 확인
         //Todo: 라운드가 시작되면 RoundProgressActivity로 전환
 
     }
 
-    fun setRecyclerView(view : View){
-        recentRoundUserAdapter = ParticipantAdapter()
-        recentRoundUserAdapter.add(UserModel("test","test"))
-
-        view.recycler_participated_projects_list.layoutManager = LinearLayoutManager(requireContext(), RecyclerView.VERTICAL, false)
-        view.recycler_participated_projects_list.adapter = recentRoundUserAdapter
-    }
-
-    fun showUserList(){
-        retrofitClient = RetrofitClient.create(InterfaceRoundUser::class.java)
-
-        retrofitClient.showRoundUser(1).enqueue(object : Callback<ResponseProjectUserListModel>{
-            override fun onFailure(call: Call<ResponseProjectUserListModel>, t: Throwable) {
-                Log.d("유저정보 통신실패", "${t}")
-            }
-
-            override fun onResponse(
-                call: Call<ResponseProjectUserListModel>,
-                response: Response<ResponseProjectUserListModel>
-            ) {
-                if(response.isSuccessful){
-                    if (response.body()!!.success){
-
-                        val data = response.body()!!.data
-                        Log.e("data",response.body()!!.toString())
-
-                        //Todo:리사이클러뷰 연결해서 뿌려주어야함
-
-                        recentRoundUserAdapter.addAll(data)
-                    }
-                }
-
-
-            }
-
-        })
-
+    //라운드 정보를 받고 나면 실행될 콜백
+    override fun afterGettingRoundInfo(roundIdx: Int) {
+        //라운드 참여
+        enterRound(roundIdx)
     }
 
     fun getUserList(){
@@ -104,9 +63,27 @@ class MemberWaitingFragment : BaseWaitingFragment(R.layout.fragment_round_settin
         SocketClient.responseEvent("roundcomplete", Emitter.Listener {
             Log.d("소켓 성공", it.toString())
 
-            showUserList()
+            //showUserList()
         })
     }
 
+    //라운드 참여
+    private fun enterRound(roundIdx: Int){
+        RetrofitClient.create(InterfaceRoundEnter::class.java).interfaceRoundEnter((RoundEnterModel(preference.getUserIdx()!!, roundIdx)))
+            .enqueue(object  : Callback<BaseResponse> {
 
+                override fun onFailure(call: Call<BaseResponse>, t: Throwable) {
+                    Log.d("라운드 참여 통신 실패", "${t}")
+                }
+
+                override fun onResponse(call: Call<BaseResponse>, response: Response<BaseResponse>) {
+                    Log.d("멤버 라운드 참여 성공", "user : ${preference.getUserIdx()}, round : $roundIdx")
+
+                    //라운드의 유저 정보 띄우기
+                    showRoundUserLIst(roundIdx)
+
+                    //Todo: 라운드기 시작되는지 소켓 통신
+                }
+            })
+    }
 }
