+++ app/src/main/res/layout/fragment_round_setting_waiting_member.xml
@@ -146,7 +146,7 @@
             </androidx.cardview.widget.CardView>
 
             <TextView
-                android:id="@+id/round_announce1"
+                android:id="@+id/round_subject"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="16dp"
@@ -158,7 +158,7 @@
                 app:layout_constraintTop_toBottomOf="@+id/cardview_roundwaiting_rulereminder" />
 
             <TextView
-                android:id="@+id/round_announce2"
+                android:id="@+id/round_time"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="ROUND 정보를 확인할 수 있습니다."
@@ -166,7 +166,7 @@
                 android:textSize="12sp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/round_announce1" />
+                app:layout_constraintTop_toBottomOf="@+id/round_subject" />
 
 
         </androidx.constraintlayout.widget.ConstraintLayout>
